users can now only grant roles to others which they currently have. Configurable by feature flag.
Added controller-level stripping of any roles that the current_user cannot grant
when adding or updating a user's memberships, obeying the feature flag.

Had to do this at controller level rather than model validations, as it makes it
extremely fiddly to set up test data, which leads to brittle tests and tech debt.